Building a Software Chain of Custody
A Guide for CTOs, CIOs, and Enterprise DevOps Teams

Executive Summary

In today’s software-driven world, organizational leaders cannot ignore the fact that hackers, viruses, malware, data breaches, and compliance violations all threaten the integrity of their software assets. Compromised software integrity can lead to operational shutdowns, contract breaches, lawsuits, and huge fines that affect revenue and profitability, damage corporate credibility, and, ultimately, cause irreparable harm to brand reputation. It’s nearly impossible to quantify the longterm cost of the consumer and investor trust that is lost after an organization is in the news for a software hack or data breach.

A Software Chain of Custody is key for proving the integrity of software assets throughout the enterprise. In the legal world, the chain of custody for a piece of evidence ensures the integrity of that evidence; similarly, the Software Chain of Custody proves what happened, when it happened, where it happened, and who made it happen during the software delivery process—from the time you plan
features through the time they’re deployed in production. A Software Chain of Custody also feeds valuable contextual data into Value Stream Management, so leaders can analyze and continuously improve software delivery processes.